+++ R/hasConverged.R
@@ -1,3 +1,4 @@
+# returns a Boolean to indicate whether the model has converged
 hasConverged <- function(model)
 {
   if (inherits(model, "MxModel") || inherits(model, "MxRAMModel")) {
@@ -5,7 +6,7 @@ hasConverged <- function(model)
     return(model$output$status$code == 0)
     
   } else if (inherits(model, "lavaan")) {
-    stop("hasConverged() is not implemented for this model class!")    
+    return(lavaan::lavInspect(model,"converged"))
   } else {
     stop("hasConverged() is not implemented for this model class!")
   }