+++ tests/testthat/test-semforest-focus-bivariate-2.R
@@ -46,7 +46,7 @@ forest <- semforest(model=model.biv, data=datf,
                     constraints = semtree.constraints(focus.parameters="COV_x1_x2"),
                     control = 
                       semforest_control(num.trees = 30,
-                                        control=semtree_control(min.N = 50, min.bucket = 100,alpha=1,
+                                        control=semtree_control(min.N = 100, min.bucket = 50,alpha=1,
                                                                 method="score")))
 
 vim_naive <- varimp(forest)